requirements.txt
djangorestframework==3.13.1
Django==4.0.4